requirements.txt
fastapi
uvicorn
pyyaml
pandas
streamlit
requests
